Robotarium/requirements.txt
absl-py==2.0.0
appdirs==1.4.4
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
Farama-Notifications==0.0.4
filelock==3.12.4
fiona==1.9.5
fonttools==4.43.1
fsspec==2023.9.2
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gym==0.25.2
gym-notices==0.0.8
gymnasium==0.28.1
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
jax-jumpy==1.0.0
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Panda3D==1.10.13
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==2.1.1
pathtools==0.1.2
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
progressbar==2.5
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
pyparsing==3.1.1
pyproj==3.6.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
sentry-sdk==1.32.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wandb==0.15.12
Werkzeug==3.0.0
yapf==0.40.2
zipp==3.17.0
